% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapsisBridgeAccessories.R
\name{convertDataSet}
\alias{convertDataSet}
\title{Convert dataset to an R equivalent}
\usage{
convertDataSet(dataSetObject)
}
\arguments{
\item{dataSetObject}{the dataSetObject java object to convert}
}
\value{
an R equivalent of the dataset
}
\description{
Convert dataset to an R equivalent
}
