% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapsisBridgeAccessories.R
\name{convertPredictions}
\alias{convertPredictions}
\title{Convert predictions returned by MMGetPredictions to an R dataframe}
\usage{
convertPredictions(pred, key)
}
\arguments{
\item{pred}{the predictions returned by MMGetPredictions}

\item{key}{string representing the key of the desired map inside (see MetaModel.PREDICTIONS)}
}
\value{
an R equivalent of the dataset
}
\description{
Convert predictions returned by MMGetPredictions to an R dataframe
}
