% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionUtilities.R
\name{.checkIfProperlyConnectedToJ4R}
\alias{.checkIfProperlyConnectedToJ4R}
\title{Check if the package is properly connect to J4R.}
\usage{
.checkIfProperlyConnectedToJ4R()
}
\description{
To be properly connected, the J4R server must be running and must include the repicea-metamodels library in its classpath.
If so, the function returns true. If the server is running but does not include the library in its classpath, the function
stops. If the J4R server is not running the function returns false.
}
