% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionUtilities.R
\name{.connectToJ4R}
\alias{.connectToJ4R}
\title{Connect to J4R Server}
\usage{
.connectToJ4R()
}
\description{
Check if the server is running and includes the repicea-metamodels library. If not running,
the server is started.
}
