% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVSClass.R
\docType{data}
\name{FVSTreeListTwoTreesOnePlot}
\alias{FVSTreeListTwoTreesOnePlot}
\title{A plot with two trees that can be passed to FVS Web API}
\format{
An object of class \code{data.frame} with 2 rows and 6 columns.
}
\usage{
data(FVSTreeListTwoTreesOnePlot)
}
\description{
A plot with two trees that can be passed to FVS Web API
}
\examples{
data(FVSTreeListTwoTreesOnePlot)
}
\keyword{datasets}
