% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapsisClass.R
\name{new_CapsisClass}
\alias{new_CapsisClass}
\title{Constructor for the CapsisClass class.}
\usage{
new_CapsisClass(host = "http://repicea.dynu.net")
}
\arguments{
\item{host}{the url of CAPSIS Web API (http://repicea.dynu.net by default)}
}
\value{
an S3 CapsisClass instance
}
\description{
This class is the interface to the CAPSIS WEB API.
}
\details{
The class contains the following methods: \cr
\itemize{

\item \bold{Simulate(data, outputRequestList, variant, years, initialYear, isStochastic, nbRealizations, climateChange, applicationScale)} \cr
Converts the current OutputRequestList to a json string \cr
\itemize{
\item data - a string in CSV format that represents the input data to run the simulation on
\item outputRequestList - An object of type OutputRequestList that contains the output data requested
\item variant - A string containing the variant name to use for simulation
\item years - An int containing the number of years the simulation should use
\item initialYear - The initial date of the simulation
\item isStochastic - A logical (true to enable stochastic simulation)
\item nbRealizations - An int containing the number of realizations in case of stochastic simulation [1-100]
\item climateChange - A string indicating the climate change scenarion [NO_CHANGE, RCP2_6, RCP4_5, RCP6_0, RCP8_5]
\item applicationScale - A string indicating the scale of the simulation [Stand, FMU]
}

}
}
