% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVSClass.R
\name{new_FVSClass}
\alias{new_FVSClass}
\title{Constructor for the FVSClass class.}
\usage{
new_FVSClass(host = "https://repicea.dynu.net")
}
\arguments{
\item{host}{the url of FVS Web API (http://repicea.dynu.net by default)}
}
\value{
an S3 FVSClass instance

 TODO update documentation here
}
\description{
This class is the interface to the OSM http server.
}
\details{
The class contains the following methods: \cr
\itemize{

\item \bold{ConvertDataFrameToCSVString(dataFrameInstance)} \cr
Utility method to convert a dataframe to CSV string to be sent as input data to the server \cr
\item dataFrameInstance - The status class to be requested (character, typically one of "Alive", "Dead" )
\item variable - The variable to be requested (character, typically one of "Volume", "Biomass")
\item aggregrationPatterns - A list of aggregation patterns to be used for the request (named List where the names are the aggregation groups, and the list data are the species)


\item \bold{Simulate(data, outputRequestList, variant, years, ypc)} \cr
Converts the current OSMOutputRequestList to a json string \cr
\item data - a string in CSV format that represents the input data to run the simulation on
\item outputRequestList - An object of type OSMOutputRequestList that contains the output data requested
\item variant - A string containing the variant name to use for simulation
\item years - An int containing the number of years the simulation should use
\item ypc - An int containing the number of years per cycle the simulation should use
Return character vector
}
}
