% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModelClass.R
\name{new_MetaModel}
\alias{new_MetaModel}
\title{Constructor for the MetaModel class.}
\usage{
new_MetaModel(stratumGroup, geoDomain, dataSource)
}
\arguments{
\item{stratumGroup}{- a descriptor of the population (e.g. RE3 potential
vegetation)}

\item{geoDomain}{- a descriptor of the location (e.g. 3OUEST)}

\item{dataSource}{- adescriptor of the data source (e.g. PET4)}
}
\value{
an S3 MetaModel instance
}
\description{
A MetaModel instance contains a pointer to a Java instance. It
also implements many methods which are described below.
}
\details{
The class contains the following methods: \cr
\itemize{

\item \bold{getStratumGroup()} \cr
Provide the stratum group the meta-model was fitted to \cr
Return a character string

\item \bold{hasConverged()} \cr
Check if the meta-model has comverged. \cr
Return a logical

\item \bold{addScriptResult(initialAge, scriptResult)} \cr
Add a simulation script result to the meta-model. Those are the simulation result
particular model. \cr
Arguments are \cr
\itemize{
\item initialDateYear - The initial year to be used for simulation (numeric)
\item scriptResult - A SimulationResult object (see the SimulationResult class)
}
Return nothing

\item \bold{getPossibleOutputTypes()} \cr
Provide the output type of the simulation. That is the possible dependent
variables of an eventual meta-model. \cr
Return a vector of character strings

\item \bold{fitModel(outputType, startingValuesMap,
randomGridSize = 10000,
nbBurnIn = 10000,
nbAcceptedRealizations = 500000 + nbBurnIn,
oneEach = 50)} \cr
Fit a metamodel to a particular output type. \cr
Arguments are \cr
\itemize{
\item outputType - The dependent variable of the meta-model
\item startingValuesMap - A StartingValuesMap instance
\item randomGridSize - The number of random trial in order to find the
starting values for the parameters (0 to disable the grid)
\item nbBurnIn - The number of burn-in realizations
\item nbAcceptedRealizations - The number of realizations in the chain
before filtering for the final sample
\item onEach - The selection rate for the final sample
}
Provide the parameter estimates in the console

\item \bold{getPredictions(ageYr, varianceOutputType, timeSinceInitialDateYr)} \cr
Provide predictions of the meta-model. \cr
Arguments are \cr
\itemize{
\item ageYr - a vector of positive integers
\item varianceOutputType - a string either NONE, PARAMEST (error on the mean),
PARAMESTRE (error on the mean + random effect)
\item timeSinceInitialDateYr - Any value (is useless at the moment, 0 by default)
}
Return a data.frame object

\item \bold{getMonteCarloPredictions(ageYr, nbSubjects, nbRealizations,
timeSinceInitialDateYr)} \cr
Provide stochastic predictions of the meta-model. \cr
Arguments are \cr
\itemize{
\item ageYr - a vector of positive integers
\item nbSubjects - the number of subject (typically strata)
\item nbRealizations - the number of realizations
\item timeSinceInitialDateYr - Any value (is useless at the moment, 0 by default)
}
Return a data.frame object

\item \bold{getSelectedOutputType()} \cr
Provide the output type of this meta-model. \cr
Return a character string

\item \bold{load(filename)} \cr
Load a meta-model from file. \cr
Arguments are \cr
\itemize{
\item filename - The name of the file that contains the meta-model instance
}
Return nothing

\item \bold{save(filename)} \cr
Save a meta-model to file. \cr
Arguments are \cr
\itemize{
\item filename - The name of the file that will contain the meta-model instance
}
Return nothing

\item \bold{getSummary} \cr
Display the parameter estimates and other information on the meta-model in
the console. \cr
Return nothing

\item \bold{getModelComparison} \cr
Display the comparison between the different meta-model implementations in
the console. \cr
Return nothing

\item \bold{plotOutputType(outputType, textsize = 20)} \cr
Plot a graph with the simulation results of a particular output type.
Arguments are \cr
\itemize{
\item outputType - One of the output types (see getPossibleOutputTypes())
\item textsize - The font size (by default 20)
\item title - The title of the graph, will use the outputType by default
}
Return nothing

\item \bold{plotFit} \cr
Provide a graph of the goodness of fit of the meta-model. \cr
Return a ggplot2 graph

\item \bold{convertToLightVersion(filename)} \cr
Produce and save a light version of the meta-model. \cr
Arguments are \cr
\itemize{
\item filename - The name of the file that will contain the meta-model instance
}
Return nothing

\item \bold{plotChain} \cr
Provide a graph of the loglikelihood of the different realizations of the
final sample. \cr
Return a ggplot2 graph

\item \bold{plotParameterEstimates} \cr
Provide a histogram for each parameter estimate. \cr
Return a list of ggplot2 graph

\item \bold{getImplementationList} \cr
Provide the list of possible implementation for the metamodels. \cr
Return a vector of characters

\item \bold{getRegenerationLagYrIfAny} \cr
Provide the regeneration lag if there is one. \cr
Return the lag (yrs)
}
}
