% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationResultClass.R
\name{new_SimulationResult}
\alias{new_SimulationResult}
\title{Constructor for the SimulationResult class.}
\usage{
new_SimulationResult(resultJSON)
}
\arguments{
\item{resultJSON}{a JSON result from a simulate call}
}
\value{
an S3 SimulationResult instance
}
\description{
This class converts the JSON result of a simulate call
into a SimulationResult instance
}
\details{
The constructor automatically detects whether the
simulation results need to be aggregated.
}
