% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModelClass.R
\name{new_StartingValues}
\alias{new_StartingValues}
\title{Constructor for the StartingValues class.}
\usage{
new_StartingValues(Parameter, StartingValue, Distribution, DistParms)
}
\arguments{
\item{Parameter}{- a vector of characters that stands for the parameter names}

\item{StartingValue}{- a vector of numerics}

\item{Distribution}{- a vector of characters that stands for the prior distributions}

\item{DistParms}{- a vector of lists containing the parameters of the prior distributions}
}
\value{
an S3 StartingValues instance
}
\description{
The parameter starting values for a particular model implementation.
}
\details{
The class contains the following methods: \cr
\itemize{

\item \bold{toJSONString()} \cr
Provide a JSON representation of this object \cr
Return a character string
}
}
