% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModelClass.R
\name{new_StartingValuesMap}
\alias{new_StartingValuesMap}
\title{Constructor for the StartingValuesMap class.}
\usage{
new_StartingValuesMap()
}
\value{
an S3 StartingValuesMap instance
}
\description{
The parameter starting values for a set of model implementations.
}
\details{
The class contains the following methods: \cr
\itemize{

\item \bold{add(implementationName, startingValues)} \cr
Add the starting values for a particular implementation. \cr
Arguments are \cr
\itemize{
\item implementationName - The implementation name (see metaModel$getImplementationList)
\item startingValues - A StartingValues instance
}
Return nothing
}
}
