% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMClass.R
\docType{data}
\name{OSMThreeStandList}
\alias{OSMThreeStandList}
\title{A list of three plots that can be passed to OSM Web API}
\format{
An object of class \code{data.frame} with 113 rows and 10 columns.
}
\usage{
data(OSMThreeStandList)
}
\description{
A list of three plots that can be passed to OSM Web API
}
\examples{
data(OSMThreeStandList)
}
\keyword{datasets}
